Move OverlayBase into Content
This allows to hide the overlay when it isn't needed.
