+++ src/v2/views/ClaimCollectionRewardsOverlay/ClaimContent.tsx
@@ -3,15 +3,15 @@ import { observer } from "mobx-react";
 import { GetAvatarAddressQuery } from "src/v2/generated/graphql";
 import { useTx } from "src/v2/utils/useTx";
 
-import type { ClaimCollectionRewardsData } from ".";
-import H1 from "src/v2/components/ui/H1";
+import type { ClaimCollectionRewardsOverlayProps } from ".";
 import { Select, SelectOption } from "src/v2/components/ui/Select";
 import Button, { ButtonBar } from "src/v2/components/ui/Button";
 import { T } from "src/renderer/i18n";
 import { styled } from "src/v2/stitches.config";
 import { getRemain } from "src/collection/common/utils";
+import OverlayBase from "src/v2/components/core/OverlayBase";
 
-interface ClaimContentProps extends ClaimCollectionRewardsData {
+interface ClaimContentProps extends ClaimCollectionRewardsOverlayProps {
   data: GetAvatarAddressQuery;
 }
 
@@ -21,7 +21,27 @@ const Title = styled("h2", {
   textAlign: "center",
 });
 
-function ClaimContent({ data, onActionTxId, rewards, tip }: ClaimContentProps) {
+const ClaimCollectionRewardsOverlayBase = styled(OverlayBase, {
+  "&&": {
+    width: 570,
+    height: 460,
+    margin: "20vh auto",
+  },
+  display: "flex",
+  flexDirection: "column",
+  "& > * + *": {
+    marginTop: 16,
+  },
+});
+
+function ClaimContent({
+  data,
+  onActionTxId,
+  rewards,
+  tip,
+  onClose,
+  isOpen,
+}: ClaimContentProps) {
   const avatars = useMemo(
     () =>
       data.stateQuery.agent?.avatarStates?.map((x) => ({
@@ -42,14 +62,14 @@ function ClaimContent({ data, onActionTxId, rewards, tip }: ClaimContentProps) {
   );
 
   useEffect(() => {
-    if (hasMultipleAvatars) return;
+    if (hasMultipleAvatars || !isOpen) return;
     tx().then((v) => v.data != null && onActionTxId(v.data.stageTxV2));
-  }, [avatars]);
+  }, [avatars, isOpen]);
 
   if (!hasMultipleAvatars) return null;
 
   return (
-    <>
+    <ClaimCollectionRewardsOverlayBase isOpen={isOpen} onDismiss={onClose}>
       <Title>
         <T
           _str="Choose a character to receive rewards."
@@ -80,7 +100,7 @@ function ClaimContent({ data, onActionTxId, rewards, tip }: ClaimContentProps) {
           <T _str="Send" _tags={transifexTags} />
         </Button>
       </ButtonBar>
-    </>
+    </ClaimCollectionRewardsOverlayBase>
   );
 }
